Refactor of database backend
This removes all the redundant insert*Info and load*Info functions
within db and the sqldb implentation (so instead, most of these
functions will try to instead call _loadEntity or _insertEntity).

This reduction in redundancy actually then allows us remove a large
number of alter*Info, insert*Info, and load*Info functions in sqldb
since they all just call the respective _*Entity function which is of
course overriden in our sqldb

+++ CreateTables.sql
@@ -91,7 +91,7 @@ create table Project (
 create table Team (
 	tid int,
   assignedProj int NULL UNIQUE,
-  budget int NOT NULL,
+  budget float NOT NULL,
   leader int,
   password varchar(100),
   PRIMARY KEY (tid),
